refactor: 인수 테스트로 변경
