+++ src/test/java/roomescape/util/JdbcConnectionTest.java
@@ -1,21 +1,22 @@
-package roomescape;
+package roomescape.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.lang.reflect.Field;
 import java.sql.Connection;
 import java.sql.SQLException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.annotation.DirtiesContext;
+import roomescape.reservation.acceptance.ReservationController;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class JdbcConnectionTest {
+public class JdbcConnectionTest extends AcceptanceTest {
     @Autowired
     private JdbcTemplate jdbcTemplate;
+
+    @Autowired
+    ReservationController reservationController;
 
     @DisplayName("DataSource 객체를 이용하여 Connection 검증한다.")
     @Test
@@ -27,5 +28,19 @@ public class JdbcConnectionTest {
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    @Test
+    void injectionTest() {
+        boolean isJdbcTemplateInjected = false;
+
+        for (Field field : reservationController.getClass().getDeclaredFields()) {
+            if (field.getType().equals(JdbcTemplate.class)) {
+                isJdbcTemplateInjected = true;
+                break;
+            }
+        }
+
+        assertThat(isJdbcTemplateInjected).isFalse();
     }
 }
